requirements.txt
jwt~=1.3.1
rel~=0.4.8
openai==0.27.4
sanic==23.3.0
aiohttp~=3.8.4
pycryptodomex==3.17
pymysql==1.1.0
sanic_cors==2.2.0
redis~=3.5.3
web3~=6.2.0
pandas~=1.5.3
setuptools~=67.6.1
bcrypt~=4.0.1
numpy~=1.24.2
PyJWT~=2.8.0
aiosmtplib~=2.0.2
Pillow~=9.5.0
captcha~=0.4
beautifulsoup4~=4.12.2
aiocron~=1.8
aiofiles~=23.1.0
scikit-learn~=1.3.0
APScheduler~=3.10.1
tradingview_ta
python-dotenv==1.0.0
qdrant-client==1.4.0
requests==2.31.0
urllib3==1.26.16
asyncio
